test_platform_agent_with_rsn re-enabled! This was after a refactor of the platform integration test set-up (initially coded in test_platform_launch) so it can be used in other test files.
